pyproject.toml
[tool.poetry]
name = "kaptilo"
version = "0.1.0"
description = "Back for kaptilo project"
authors = ["DrMartiner"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
psycopg2-binary = "^2.8"
gunicorn = "^19.9"
django-environ = "^0.4.5"
dj-database-url = "^0.5.0"
django-ipware = "^3.0.2"
django-widget-tweaks = "^1.4.8"
sqlparse = "^0.4.1"
chardet = "^4.0.0"
django-admin-honeypot = "^1.1.0"
django-rest-framework = "^0.1.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = {version = "^19.10b0", allow-prereleases = true}
isort = "^4.3"
django-dynamic-fixture = "^3.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
